requirements.txt
torchvision==0.14.1
torchaudio==0.13.1
numpy>=1.21.5
PyYAML>=5.4.1
networkx>=2.8.4
Cython
notebook
timm
local-attention
einops
